refactor: 컨트롤러 기능을 수행하는 패키지를 명확한 이름으로 변경
